requirements.txt
flask==2.3.3
zappa
openai==0.28.1
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0 